restructured the preload predictor
