+++ sample.cfg
@@ -2,7 +2,7 @@
 # Each module that should be used in the simulator
 # should be specified in the modules field. The
 # modules listed should be separated by spaces.
-modules = preload-predictor
+modules = chrome-preload-predictor
 
 # Other simulator specific settings should also
 # be defined here
