+++ test.csv
@@ -1,4 +1,4 @@
-app,overall,mor,non,ngt
-google maps,9.999999997693868,0,0,9.999999997693868
-facebook,9.99999999771752,0,0,9.99999999771752
-emails,99.9999999796469,0,0,99.9999999796469
+app,overall,mor,non,ngt
+google maps,9.999999302179237,0,9.999999302179237,0
+facebook,9.999999290060217,0,9.999999290060217,0
+emails,99.99999290060215,0,99.99999290060215,0
